Move a few views to the new views module (#423)
* wip

* wip

* clean up

* clean up
