+++ Sources/Views/TODO.swift
@@ -1,0 +1,5 @@
+import Html
+
+public func playsinline(_ value: Bool) -> Attribute<Tag.Video> {
+  return .init("playslinline", value ? "" : nil)
+}
